web/requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-HTTPAuth==4.4.0
flask-profiler==1.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.12.0
simplejson==3.17.5
Werkzeug==2.0.1
python-docx==0.8.11
